Abstract out repo link input for pushing
